# greengrocers-products-catalogue/src/main/resources/schemas/products-read-api-spec.yaml
openapi: '3.0.3'
info:
  title: Products inventory API for Green Grocers
  version: '1.0.0'
tags:
  - name: admins
    description: Administrative actions for suppliers data
  - name: developers
    description: Development API for suppliers and products data
servers:
  - url: https://greengrocer/content/v1
paths:
  /content/v1/health-check:
    get:
      tags:
        - developers
      summary: Health check ping
      operationId: healthCheck
      description: Perform health check of product inventory services availability
      responses:
        '200':
          description: All good!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckStatus'
        '400':
          description: System unhealthy; find a solution to avoid disrupting others!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckStatus'

  /content/v1/suppliers:
    get:
      tags:
        - admins
      summary: List of all available suppliers
      operationId: suppliers
      description: |
        Fetch all the suppliers of greengrocer (admin only)
      responses:
        '200':
          description: Return the list of all suppliers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierSummary'

  /content/v1/suppliers/{supplierId}/products:
    get:
      tags:
        - developers
      summary: List of all available products for the given supplier
      operationId: suppliersProducts
      description: |
        Fetch all products supplied by a specific suppliers (developers only)
      parameters:
        - in: path
          name: supplierId
          required: true
          schema:
            type: string
            example: "123"
      responses:
        '200':
          description: List of products by the supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsSummary'


  /content/v1/suppliers/{supplierId}/products/{productId}:
    get:
      tags:
        - developers
      summary: Details of a product
      operationId: product
      description: Gets details a product
      parameters:
        - in: path
          name: supplierId
          schema:
            type: string
            example: "123"
          required: true
        - in: path
          name: productId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Details of the product

  /content/v1/products:
    get:
      tags:
        - developers
      summary: All product
      operationId: products
      description: Gets a list of all products
      responses:
        '200':
          description: List of all products

components:
  schemas:
    HealthCheckStatus:
      type: string

    SupplierSummary:
      type: array
      items:
        type: object
        required:
          - id
          - name
        properties:
          id:
            type: string
            example: "UK-123"
          name:
            type: string
            example: "World Record Breaker"
          products:
            type: array
            items:
              $ref: '#/components/schemas/ProductsSummary'

    ProductsSummary:
      type: object
      required:
        - productId
        - name
        - sellUnitPrice
        - sellUnitCurrency
        - barCode
      properties:
        productId:
          type: string
          example: "345"
        name:
          type: string
          example: "Kulkuli"
        sellUnit:
          type: string
          example: "piece"
        sellUnitPrice:
          type: number
          format: double
          example: 3.0
        sellUnitCurrency:
          type: string
          example: "GBP"
        weight:
          type: integer
          example: "250"
        weightUnit:
          type: string
          example: "ml"
        imageUrl:
          type: string
          example: "https://greengrocers/images/4466"
        catalogIdsReference:
          type: string
        barCode:
          type: string
          example: "99990"
        ingredients:
          type: array
          items:
            type: string
          example:
            - Coco
            - Sugar
            - Milk
      example:
        - productId: 345
          name: "Kulkuli"
          sellUnitPrice: 3.0
          sellUnitCurrency: "GBP"
          barCode: "99999"